# docs/contracts/http/schemaapi.openapi.yaml
openapi: 3.0.3
info:
  title: SchemaAPI
  version: 0.5.0
  description: Document Intelligence API with ingestion, hybrid retrieval, cited RAG, governance, agents, analysis reports and observability.
servers:
  - url: http://localhost:8081
paths:
...
  /agents/runs:
    post:
      summary: Create a controlled agent run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goal]
              properties:
                goal:
                  type: string
                requested_tool:
                  type: string
      responses:
        "202":
          description: Agent run created
  /agents/runs/{id}:
    get:
      summary: Get an agent run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Agent run
        "404":
          description: Agent run not found
  /agents/runs/{id}/approve:
    post:
      summary: Approve a sensitive agent run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved_by]
              properties:
                approved_by:
                  type: string
      responses:
        "200":
          description: Approved agent run
        "404":
          description: Agent run not found
...